refactor: 피스 별 MoveVerifier 테스트 정리
